Implemented saving order into database

+++ app/src/main/java/bohdan/sushchak/mywallet/internal/ListExtension.kt
@@ -2,7 +2,7 @@ package bohdan.sushchak.mywallet.internal
 
 import bohdan.sushchak.mywallet.data.db.entity.Product
 
-fun List<Product>.setOrderId(orderId: Int) {
+fun List<Product>.setOrderId(orderId: Long) {
     forEachIndexed { index, product ->
         this[index].orderId = orderId
     }
